+++ djangobmf/contrib/invoice/models.py
@@ -162,7 +162,7 @@ class Invoice(AbstractInvoice):
     pass
 
 
-class InvoiceProduct(models.Model):
+class InvoiceProduct(BMFModel):
     invoice = models.ForeignKey(
         CONTRIB_INVOICE, null=True, blank=True,
         related_name="invoice_products", on_delete=models.CASCADE,
@@ -181,18 +181,8 @@ class InvoiceProduct(models.Model):
     # unit = models.CharField() # TODO add units
     description = models.TextField(_("Description"), null=True, blank=True)
 
-    def __init__(self, *args, **kwargs):
-        super(InvoiceProduct, self).__init__(*args, **kwargs)
-        self._calcs = None
-
-    def clean(self):
-        if self.product and not self.name:
-            self.name = self.product.name
-        if self.product and not self.price:
-            self.price = self.product.price
-
     def calc_all(self):
-        if self._calcs:
+        if hasattr(self, '_calcs'):
             return self._calcs
         self._calcs = self.product.calc_tax(self.amount, self.price)
         return self._calcs
@@ -208,3 +198,9 @@ class InvoiceProduct(models.Model):
 
     def calc_taxes(self):
         return self.calc_all()[3]
+
+    def clean(self):
+        if self.product and not self.name:
+            self.name = self.product.name
+        if self.product and not self.price:
+            self.price = self.product.price